Natural Language processing/requirements_for_question1.txt
spacy
pyspellchecker